+++ x/mint/keeper/abci.go
@@ -1,11 +1,10 @@
-package mint
+package keeper
 
 import (
 	"context"
 	"time"
 
 	"cosmossdk.io/core/event"
-	"cosmossdk.io/x/mint/keeper"
 	"cosmossdk.io/x/mint/types"
 
 	"github.com/cosmos/cosmos-sdk/telemetry"
@@ -13,7 +12,7 @@ import (
 )
 
 // BeginBlocker mints new tokens for the previous block.
-func BeginBlocker(ctx context.Context, k keeper.Keeper, ic types.InflationCalculationFn) error {
+func (k Keeper) BeginBlocker(ctx context.Context, ic types.InflationCalculationFn) error {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyBeginBlocker)
 
 	// fetch stored minter & params
@@ -63,7 +62,7 @@ func BeginBlocker(ctx context.Context, k keeper.Keeper, ic types.InflationCalcul
 		defer telemetry.ModuleSetGauge(types.ModuleName, float32(mintedCoin.Amount.Int64()), "minted_tokens")
 	}
 
-	return k.EventService.EventManager(ctx).EmitKV(
+	return k.environment.EventService.EventManager(ctx).EmitKV(
 		types.EventTypeMint,
 		event.NewAttribute(types.AttributeKeyBondedRatio, bondedRatio.String()),
 		event.NewAttribute(types.AttributeKeyInflation, minter.Inflation.String()),
